+++ doc/cookbook/simd_dna4.cpp
@@ -1,0 +1,9 @@
+// Checks that defining simd_dna4 works without putting it into the seqan3 namespace.
+
+#define SEQAN3_USE_NAMESPACE 0
+#include <seqan3/test/performance/simd_dna4.hpp>
+
+int main()
+{
+    simd_dna4 [[maybe_unused]] letter{};
+}